Feat: Add VirtualKeyboard Component
