+++ src/pages/MainPage/index.css
@@ -1,9 +1,9 @@
-*{
-  font-family: 'Montserrat', sans-serif;
+* {
+  font-family: "Montserrat", sans-serif;
   color: #293677;
 }
 .container {
-  border: 1px solid #B1B2FF;
+  border: 1px solid #b1b2ff;
   margin: 0 auto;
   width: 1080px;
   height: 2000px;
@@ -17,41 +17,37 @@ header {
   width: 1040px;
   height: 50px;
 
-  border: 1px solid #B1B2FF;
-  background-color: #EEF1FF;
+  border: 1px solid #b1b2ff;
+  background-color: #eef1ff;
   position: fixed;
   top: 0;
   left: 50%;
   transform: translate(-50%);
 }
 
-
-.header_logo{
+.header_logo {
   width: 80px;
-
-
 }
 .center_container {
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: column;
-  margin: 20px;
+
   text-align: center;
-  
 }
 .start_button {
   margin: 20px;
   width: 350px;
   height: 50px;
   border-radius: 10px;
-  background-color: #B1B2FF;
+  background-color: #b1b2ff;
   border: none;
   color: #293677;
   font-size: 15pt;
 }
 
-.start_button:hover{
+.start_button:hover {
   filter: brightness(0.9);
   cursor: pointer;
 }
@@ -60,7 +56,6 @@ header {
   margin: 20px;
   padding-left: 20px;
   padding-bottom: 20px;
-  
 }
 .paragraph_title {
   padding-top: 10px;
@@ -81,23 +76,23 @@ p {
   line-height: 1.5;
 }
 
-.title{
+.title {
   font-size: 30pt;
   color: purple;
 }
 
-.logo{
+.main_page_logo {
   margin: 50px;
   width: 400px;
 }
 
-.nav{
+.nav {
   margin-top: 3px;
   display: flex;
   font-size: 10.5pt;
   justify-content: space-between;
-  width: 140px; 
+  width: 140px;
 }
-a{
+a {
   text-decoration: none;
 }
